requirements.txt
flask==3.0.1
gunicorn
werkzeug==3.0.1
python-dotenv
requests
PyPDF2
